pyproject.toml
[project]
name = "quantum-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Dev_Niel",email = "aryeeothniel@yahoo.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "qiskit (>=2.0.0,<3.0.0)",
    "qiskit-aer (>=0.17.0,<0.18.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "openai (>=1.78.1,<2.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "ipfshttpclient (>=0.7.0,<0.8.0)",
    "requests (>=2.32.3,<3.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
